useElectronMenu, name
